requirements.txt
slack_bolt
slack_sdk
python-dotenv
aiohttp
openai